# cmd/chat-bots-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Bot Factory backend API written in Go Gin",
        "title": "Bot Factory API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/chat-bot/": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Save chat-bot object to database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat-bot"
                ],
                "summary": "Saves chat-bot",
                "parameters": [
                    {
                        "description": "Chat bot",
                        "name": "chat_bot",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.SaveChatBotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.SaveChatBotResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/chat-bot/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetch chat-bot object from database by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat-bot"
                ],
                "summary": "Fetches chat-bot by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Chat-bot ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.ChatBot"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes chat-bot object from database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat-bot"
                ],
                "summary": "Removes chat-bot",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Chat bot",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.RemoveChatBotResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update chat-bot object in database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat-bot"
                ],
                "summary": "Updates chat-bot",
                "parameters": [
                    {
                        "description": "Chat bot",
                        "name": "chat_bot",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.UpdateChatBotRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Chat bot",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.UpdateChatBotResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/chat-bot/{id}/chat/ws": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Starts chat with a chat-bot with id = chatBotID, opens a websocket connection",
                "tags": [
                    "chat-bot"
                ],
                "summary": "Start chat with chat-bot",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Chat bot",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/chat-bots": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all user chat-bots from database",
                "tags": [
                    "user"
                ],
                "summary": "Get all user chat-bots",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.UserChatBots"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "schemas.ChatBot": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "instructions": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "schemas.RemoveChatBotResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "schemas.SaveChatBotRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "instructions": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "schemas.SaveChatBotResponse": {
            "type": "object",
            "properties": {
                "chat_bot_id": {
                    "type": "integer"
                }
            }
        },
...
        "schemas.UpdateChatBotRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "instructions": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "schemas.UpdateChatBotResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "schemas.UserChatBots": {
            "type": "object",
            "properties": {
                "chat_bots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.ChatBot"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "cookie"
        }
    }
}